requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
django==3.2.8
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-cors-headers==3.10.0
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   rest-framework-generic-relations
faker==9.3.1
    # via -r requirements.in
ibm-cloud-sdk-core==3.12.0
    # via ibm-watson
ibm-watson==5.3.0
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.8.1
    # via markdown
markdown==3.3.4
    # via -r requirements.in
pillow==8.3.2
    # via -r requirements.in
psycopg2==2.9.1
    # via -r requirements.in
pyjwt==2.3.0
    # via ibm-cloud-sdk-core
python-dateutil==2.8.2
    # via
    #   faker
    #   ibm-cloud-sdk-core
    #   ibm-watson
pytz==2021.3
    # via django
requests==2.26.0
    # via
    #   -r requirements.in
    #   ibm-cloud-sdk-core
    #   ibm-watson
rest-framework-generic-relations==2.1.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
text-unidecode==1.3
    # via faker
typing-extensions==3.10.0.2
    # via
    #   asgiref
    #   importlib-metadata
urllib3==1.26.7
    # via requests
websocket-client==1.1.0
    # via ibm-watson
zipp==3.6.0
    # via importlib-metadata
